+++ plugin/src/main/groovy/org/openbakery/appstore/AppstorePluginExtension.groovy
@@ -1,0 +1,15 @@
+package org.openbakery.appstore
+
+import org.gradle.api.Project
+
+/**
+ * Created by rene on 08.01.15.
+ */
+class AppstorePluginExtension {
+
+	def String username=null
+	def String password=null
+
+	public AppstorePluginExtension(Project project) {
+	}
+}