+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.10.0 (TAB)
 
 Changes
+- Renamed all tasks with dashes in the name and removed the dash, e.g hockeykit-clean to hockeykitClean
 - New package task to creates IPAs
   - supports Swift
 	- supports App Extensions
@@ -9,6 +10,8 @@ Changes
 - archive task now creates a xcarchive: Issue #1
 - distribution tasks (e.g. hockeykit) does not have a dependences to the build tasks anymore, if you want this define this in you build.gradle file
 - distribution tasks uses the xcarchive as input now
+- Added new 'appstore-upload' target that submits the ipa to apple
+- Remove Testflight. Is Replaced by the appstore-upload target
 
 
 ## 0.9.15 (November 25, 2014)